+++ confluence/legacy-mobile-macros/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/legacy-mobile-macros
 
+## 2.4.89
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 2.4.88
 
 ### Patch Changes