+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-mobile-bridge
 
+## 187.10.5
+
+### Patch Changes
+
+- Updated dependencies
+
+## 187.10.4
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 187.10.3
 
 ### Patch Changes